refactor(dirs): restructure subpackage layout and APIs
